adding DTO class to the controller

+++ src/main/java/cat/itacademy/barcelonactiva/jimenez/leonardo/s05/t02/n01/controllers/PlayerController.java
@@ -2,6 +2,7 @@ package cat.itacademy.barcelonactiva.jimenez.leonardo.s05.t02.n01.controllers;
 
 
 import cat.itacademy.barcelonactiva.jimenez.leonardo.s05.t02.n01.model.domain.Player;
+import cat.itacademy.barcelonactiva.jimenez.leonardo.s05.t02.n01.model.dto.PlayerDTO;
 import cat.itacademy.barcelonactiva.jimenez.leonardo.s05.t02.n01.model.services.PlayerService;
 import ch.qos.logback.classic.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,8 +11,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/players")
@@ -22,11 +25,11 @@ public class PlayerController {
     PlayerService playerService;
 
     @PostMapping("/add")
-    public ResponseEntity<Player> add(@RequestBody Player player) {
+    public ResponseEntity<PlayerDTO> add(@RequestBody PlayerDTO playerDTO) {
         logger.info("Calling add method");
         try {
-            playerService.add(player);
-            return new ResponseEntity<>(player, HttpStatus.CREATED);
+            Player player = playerService.add(playerService.convertToEntity(playerDTO));
+            return new ResponseEntity<>(playerService.convertToDto(player), HttpStatus.CREATED);
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -34,14 +37,16 @@ public class PlayerController {
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Player> update(@PathVariable("id") long id, @RequestBody Player playerDetails) {
+    public ResponseEntity<PlayerDTO> update(@PathVariable("id") long id, @RequestBody PlayerDTO playerDetails) {
         logger.info("Calling update method");
         try {
             Optional<Player> player = playerService.findById(id);
-            if (player.isPresent()) {
+            if (id > 0 && player.isPresent()) {
                 player.get().setName(playerDetails.getName());
+                player.get().setRegistrationDate(new Date());
                 playerService.update(player.get());
-                return new ResponseEntity<>(playerService.findById(id).get(), HttpStatus.OK);
+                return new ResponseEntity<>(playerService.convertToDto(playerService.findById(id).get()),
+                        HttpStatus.OK);
             }
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
@@ -51,11 +56,11 @@ public class PlayerController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Player> delete(@PathVariable("id") long id) {
+    public ResponseEntity<PlayerDTO> delete(@PathVariable("id") long id) {
         logger.info("Calling delete method");
         try {
             Optional<Player> player = playerService.findById(id);
-            if (player.isPresent()) {
+            if (id > 0 && player.isPresent()) {
                 playerService.deleteById(id);
                 return new ResponseEntity<>(HttpStatus.OK);
             }
@@ -67,13 +72,14 @@ public class PlayerController {
     }
 
     @GetMapping("/getOne/{id}")
-    public ResponseEntity<Player> getOne(@PathVariable("id") long id) {
+    public ResponseEntity<PlayerDTO> getOne(@PathVariable("id") long id) {
         logger.info("Calling getOne method");
         try {
             Optional<Player> player = playerService.findById(id);
-            if (player.isPresent()) {
-                playerService.deleteById(id);
-                return new ResponseEntity<>(player.get(), HttpStatus.OK);
+            if (id > 0 && player.isPresent()) {
+                playerService.findById(id);
+                return new ResponseEntity<>(playerService.convertToDto(playerService.findById(id).get()),
+                        HttpStatus.OK);
             }
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         } catch (Exception e) {
@@ -82,11 +88,11 @@ public class PlayerController {
     }
 
     @GetMapping("/getAll")
-    public ResponseEntity<List<Player>> getAll() {
+    public ResponseEntity<List<PlayerDTO>> getAll() {
         logger.info("Calling getAll method");
         try {
-            List<Player> fruites = playerService.getAll();
-            return new ResponseEntity<>(fruites, HttpStatus.OK);
+            return new ResponseEntity<>(playerService.getAll().stream().map(
+                    s -> playerService.convertToDto(s)).collect(Collectors.toList()), HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
